requirements.txt
Django==2.2
boto3==1.9.134
botocore==1.12.134
djangorestframework==3.9.2
docutils==0.14
jmespath==0.9.4
pip==19.0.3
python-dateutil==2.8.0
pytz==2019.1
s3transfer==0.2.0
setuptools==40.8.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.2
wheel==0.33.1